+++ CHANGELOG.md
@@ -1,12 +1,19 @@
-# Staged
+# PENDING
+
+* Fix standalone cta button [#7436](https://github.com/Applifier/unity-ads-webview/pull/7436)
+
+# Wed Dec 12 10:27:59 UTC 2018
 
 * Use env.revision for commit id [#7510](https://github.com/Applifier/unity-ads-webview/pull/7510)
+* Log Promo Error whenever creatives not found [#7465](https://github.com/Applifier/unity-ads-webview/pull/7465)
+* LoopMe MRAID impression count fix experiment [#7468](https://github.com/Applifier/unity-ads-webview/pull/7468)
 * Add AdUnitParametersFactory to reduce clutter of Ads and AdUnitFactory [#7321](https://github.com/Applifier/unity-ads-webview/pull/7321)
 * Remove Show install now button after 0,5s experiment [#7517](https://github.com/Applifier/unity-ads-webview/pull/7517)
 * Add logic to use commit id of deployed webview [#7518](https://github.com/Applifier/unity-ads-webview/pull/7518)
 * Refactor any to unknown [#7319](https://github.com/Applifier/unity-ads-webview/pull/7319)
 * Send ad plan with configuration_ad_unit_style_parse_error diagnostic error [#7511](https://github.com/Applifier/unity-ads-webview/pull/7511)
 * Use CHANGE_BRANCH instead of BRANCH_NAME [#7540](https://github.com/Applifier/unity-ads-webview/pull/7540)
+* Revert "Ramp up Auction V5 to 45% of traffic [#7514](https://github.com/Applifier/unity-ads-webview/pull/7514)" (Back to 15% of traffic)
 
 # Mon Dec 10 22:50:56 UTC 2018